+++ decidim-proposals/app/events/decidim/proposals/accepted_proposal_event.rb
@@ -2,38 +2,8 @@
 
 module Decidim
   module Proposals
-    class AcceptedProposalEvent < Decidim::Events::BaseEvent
-      include Decidim::Events::EmailEvent
-      include Decidim::Events::NotificationEvent
-
-      def email_subject
-        I18n.t(
-          "decidim.proposals.events.accepted_proposal_event.email_subject",
-          resource_title: resource_title
-        )
-      end
-
-      def email_intro
-        I18n.t(
-          "decidim.proposals.events.accepted_proposal_event.email_intro",
-          resource_title: resource_title
-        )
-      end
-
-      def email_outro
-        I18n.t(
-          "decidim.proposals.events.accepted_proposal_event.email_outro",
-          resource_title: resource_title
-        )
-      end
-
-      def notification_title
-        I18n.t(
-          "decidim.proposals.events.accepted_proposal_event.notification_title",
-          resource_title: resource_title,
-          resource_path: resource_path
-        ).html_safe
-      end
+    class AcceptedProposalEvent < Decidim::Events::SimpleEvent
+      include Decidim::Events::AuthorEvent
     end
   end
 end